+++ RxOn/server/main.js
@@ -1,6 +1,6 @@
 import '../imports/collections/prescriptions.js';
 import '../imports/collections/server/prescription_methods.js';
-import '../imports/collections/server/accounts_methods.js';
+import '../imports/collections/server/accounts_methods_publications.js';
 import '../imports/collections/server/prescription_publications.js';
 import '../imports/startup/server/accounts-config.js';
 import '../imports/collections/server/images';
